Break up server.ts.
