+++ lsif/src/server/middleware/metrics.ts
@@ -1,0 +1,35 @@
+import express from 'express'
+import onFinished from 'on-finished'
+import promClient from 'prom-client'
+import { httpQueryDurationHistogram, httpUploadDurationHistogram } from '../../server.metrics'
+
+/**
+ * Middleware function used to emit HTTP durations for LSIF functions. Originally
+ * we used an express bundle, but that did not allow us to have different histogram
+ * bucket for different endpoints, which makes half of the metrics useless in the
+ * presence of large uploads.
+ */
+export const metricsMiddleware = (req: express.Request, res: express.Response, next: express.NextFunction): void => {
+    let histogram: promClient.Histogram | undefined
+    switch (req.path) {
+        case '/upload':
+            histogram = httpUploadDurationHistogram
+            break
+
+        case '/exists':
+        case '/request':
+            histogram = httpQueryDurationHistogram
+    }
+
+    if (histogram !== undefined) {
+        const labels = { code: 0 }
+        const end = histogram.startTimer(labels)
+
+        onFinished(res, () => {
+            labels.code = res.statusCode
+            end()
+        })
+    }
+
+    next()
+}
